requirements.txt
opencv-python
torch
numpy
ultralytics
pandas
tqdm
seaborn
pillow